requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0
pdfkit==1.0.0
markdown2==2.4.10
langdetect==1.0.9
huggingface-hub==0.21.4
Pillow==10.2.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
gunicorn==21.2.0 